requirements.txt
fastapi
uvicorn
streamlit
transformers
sentence-transformers
faiss-cpu
requests
torch
